coupon has usded or not tracked in ActivitiesRecordHistory

+++ src/app/modules/activitiesRecordHistory/activitiesRecordHistory.service.ts
@@ -3,79 +3,60 @@ import AppError from '../../../errors/AppError';
 import { IactivitiesRecordHistory } from './activitiesRecordHistory.interface';
 import { ActivitiesRecordHistory } from './activitiesRecordHistory.model';
 import QueryBuilder from '../../builder/QueryBuilder';
-import unlinkFile from '../../../shared/unlinkFile';
+import { IJwtPayload } from '../auth/auth.interface';
 
-const createActivitiesRecordHistory = async (payload: IactivitiesRecordHistory): Promise<IactivitiesRecordHistory> => {
-     const result = await ActivitiesRecordHistory.create(payload);
+const createActivitiesRecordHistory = async (payload: IactivitiesRecordHistory, user: IJwtPayload): Promise<IactivitiesRecordHistory> => {
+     const isExistRecord = await ActivitiesRecordHistory.findOne({
+          moduleDocumentId: payload.moduleDocumentId,
+          historyOfModule: payload.historyOfModule,
+          userId: user.id,
+     });
+
+     if (isExistRecord) {
+          return isExistRecord;
+     }
+     const payloadData = { ...payload, userId: user.id };
+     const result = await ActivitiesRecordHistory.create(payloadData);
      if (!result) {
-          if(payload.image){
-               unlinkFile(payload.image);
-          }
           throw new AppError(StatusCodes.NOT_FOUND, 'ActivitiesRecordHistory not found.');
      }
      return result;
 };
 
-const getAllActivitiesRecordHistorys = async (query: Record<string, any>): Promise<{ meta: { total: number; page: number; limit: number; }; result: IactivitiesRecordHistory[]; }> => {
-     const queryBuilder = new QueryBuilder(ActivitiesRecordHistory.find(), query);
+const getAllActivitiesRecordHistorys = async (query: Record<string, any>): Promise<{ meta: { total: number; page: number; limit: number }; result: IactivitiesRecordHistory[] }> => {
+     console.log('first');
+     const queryBuilder = new QueryBuilder(ActivitiesRecordHistory.find().populate('moduleDocumentId'), query);
      const result = await queryBuilder.filter().sort().paginate().fields().modelQuery;
      const meta = await queryBuilder.countTotal();
      return { meta, result };
 };
 
 const getAllUnpaginatedActivitiesRecordHistorys = async (): Promise<IactivitiesRecordHistory[]> => {
-     const result = await ActivitiesRecordHistory.find();
+     const result = await ActivitiesRecordHistory.find().populate('moduleDocumentId');
      return result;
 };
 
-const updateActivitiesRecordHistory = async (id: string, payload: Partial<IactivitiesRecordHistory>): Promise<IactivitiesRecordHistory | null> => {
-     const isExist = await ActivitiesRecordHistory.findById(id);
-     if (!isExist) {
-          if(payload.image){
-               unlinkFile(payload.image);
-          }
-          throw new AppError(StatusCodes.NOT_FOUND, 'ActivitiesRecordHistory not found.');
+const hardDeleteActivitiesRecordHistory = async (id: string, user: IJwtPayload): Promise<IactivitiesRecordHistory | null> => {
+     const isExistRecord = await ActivitiesRecordHistory.findOne({ _id: id, userId: user.id });
+     if (!isExistRecord) {
+          throw new AppError(StatusCodes.NOT_FOUND, 'You are not authorized to delete this tracking record.');
      }
-
-     if(isExist.image){
-          unlinkFile(isExist.image);
-     }
-     return await ActivitiesRecordHistory.findByIdAndUpdate(id, payload, { new: true });
-};
-
-const deleteActivitiesRecordHistory = async (id: string): Promise<IactivitiesRecordHistory | null> => {
-     const result = await ActivitiesRecordHistory.findById(id);
-     if (!result) {
-          throw new AppError(StatusCodes.NOT_FOUND, 'ActivitiesRecordHistory not found.');
-     }
-     result.isDeleted = true;
-     result.deletedAt = new Date();
-     await result.save();
-     return result;
-};
-
-const hardDeleteActivitiesRecordHistory = async (id: string): Promise<IactivitiesRecordHistory | null> => {
      const result = await ActivitiesRecordHistory.findByIdAndDelete(id);
      if (!result) {
           throw new AppError(StatusCodes.NOT_FOUND, 'ActivitiesRecordHistory not found.');
      }
-     if(result.image){
-          unlinkFile(result.image);
-     }
      return result;
 };
 
-const getActivitiesRecordHistoryById = async (id: string): Promise<IactivitiesRecordHistory | null> => {
-     const result = await ActivitiesRecordHistory.findById(id);
-     return result;
-};   
+const getActivitiesRecordHistoryById = async (id: string, user: IJwtPayload): Promise<IactivitiesRecordHistory | null> => {
+     const isExit = await ActivitiesRecordHistory.findOne({ _id: id, userId: user.id }).populate('moduleDocumentId');
+     return isExit;
+};
 
 export const activitiesRecordHistoryService = {
      createActivitiesRecordHistory,
      getAllActivitiesRecordHistorys,
      getAllUnpaginatedActivitiesRecordHistorys,
-     updateActivitiesRecordHistory,
-     deleteActivitiesRecordHistory,
      hardDeleteActivitiesRecordHistory,
-     getActivitiesRecordHistoryById
+     getActivitiesRecordHistoryById,
 };
